Initial GsapCompose implementation
