+++ src/gsap/index.js
@@ -2,6 +2,5 @@ export * from './gsap.js';
 // export * from './gsapNPM.js';
 
 // export * from './animate.js';
-export * from './GsapPosition.js';
-
+export * from './compose/GsapCompose.js';
